Moved io.fromimage to utils
Since the io module will have to go, I started moving things to other
modules. fromimage will go to utils for now since I couldn't think of a
better fit.
